Add const-compatible `iovec` for `writev(2)`
`iovec` does not reflect that `writev(2)` does not mutate the data passed
to it, making it impossible to pass `const` data to `writev()`. Add
`iovec_const`, which is a duplicate of `iovec` except that `iov_base` is
declared to reference `const` data, and adjust the `writev()` declaration
to accept an array of `iovec_const`.